+++ libavcodec/jrevdct.c
@@ -943,13 +943,13 @@ void ff_j_rev_dct(DCTBLOCK data)
   }
 }
 
-void ff_jref_idct_put(uint8_t *dest, int line_size, int16_t *block)
+void ff_jref_idct_put(uint8_t *dest, ptrdiff_t line_size, int16_t *block)
 {
     ff_j_rev_dct(block);
     ff_put_pixels_clamped(block, dest, line_size);
 }
 
-void ff_jref_idct_add(uint8_t *dest, int line_size, int16_t *block)
+void ff_jref_idct_add(uint8_t *dest, ptrdiff_t line_size, int16_t *block)
 {
     ff_j_rev_dct(block);
     ff_add_pixels_clamped(block, dest, line_size);
