+++ libavcodec/x86/idctdsp.h
@@ -19,13 +19,14 @@
 #ifndef AVCODEC_X86_IDCTDSP_H
 #define AVCODEC_X86_IDCTDSP_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 void ff_add_pixels_clamped_mmx(const int16_t *block, uint8_t *pixels,
-                               int line_size);
+                               ptrdiff_t line_size);
 void ff_put_pixels_clamped_mmx(const int16_t *block, uint8_t *pixels,
-                               int line_size);
+                               ptrdiff_t line_size);
 void ff_put_signed_pixels_clamped_mmx(const int16_t *block, uint8_t *pixels,
-                                      int line_size);
+                                      ptrdiff_t line_size);
 
 #endif /* AVCODEC_X86_IDCTDSP_H */
