+++ libavcodec/x86/simple_idct.h
@@ -19,10 +19,11 @@
 #ifndef AVCODEC_X86_SIMPLE_IDCT_H
 #define AVCODEC_X86_SIMPLE_IDCT_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 void ff_simple_idct_mmx(int16_t *block);
-void ff_simple_idct_add_mmx(uint8_t *dest, int line_size, int16_t *block);
-void ff_simple_idct_put_mmx(uint8_t *dest, int line_size, int16_t *block);
+void ff_simple_idct_add_mmx(uint8_t *dest, ptrdiff_t line_size, int16_t *block);
+void ff_simple_idct_put_mmx(uint8_t *dest, ptrdiff_t line_size, int16_t *block);
 
 #endif /* AVCODEC_X86_SIMPLE_IDCT_H */
